Move appTransfer class to a separate file
